+++ src/w_resrc.h
@@ -4,13 +4,13 @@
 struct Resource
 {
 	char* buf{ nullptr };
-	long size{ 0 };
+	int size{ 0 };
 	char name[4]{};
-	long id{ 0 };
+	int id{ 0 };
 	struct Resource* next{ nullptr };
 };
 
 
 typedef char** Handle;
 
-Handle GetResource(char* name, long id);
+Handle GetResource(char* name, int id);
